deprecation: Renamed the {{local-class-new}} helper to {{local}} (embroider-css-modules-temporary)
