+++ packages/embroider-css-modules-temporary/rollup.config.mjs
@@ -15,7 +15,7 @@ export default {
     // These are the modules that users should be able to import from your
     // addon. Anything not listed here may get optimized away.
     addon.publicEntrypoints([
-      'helpers/local-class-new.js',
+      'helpers/**/*.js',
       'index.js',
       'template-registry.js',
     ]),
@@ -23,7 +23,7 @@ export default {
     // These are the modules that should get reexported into the traditional
     // "app" tree. Things in here should also be in publicEntrypoints above, but
     // not everything in publicEntrypoints necessarily needs to go here.
-    addon.appReexports(['helpers/local-class-new.js']),
+    addon.appReexports(['helpers/**/*.js']),
 
     // Follow the V2 Addon rules about dependencies. Your code can import from
     // `dependencies` and `peerDependencies` as well as standard Ember-provided
