+++ packages/embroider-css-modules-temporary/src/helpers/local-class-new.ts
@@ -1,53 +1,20 @@
-import Helper from '@ember/component/helper';
-import { assert } from '@ember/debug';
+import { deprecate } from '@ember/debug';
 
-type IndexSignatureParameter = string | number | symbol;
+import LocalHelper from './local.ts';
 
-type LocalClassName<T extends IndexSignatureParameter> = T;
+deprecate(
+  'The {{local-class-new}} helper has been renamed to {{local}}. Please update the helper name in your template.',
+  false,
+  {
+    for: 'embroider-css-modules-temporary',
+    id: 'embroider-css-modules-temporary.rename-local-class-helper',
+    since: {
+      available: '1.1.0',
+      enabled: '2.0.0',
+    },
+    until: '2.0.0',
+    url: 'https://github.com/ijlee2/embroider-css-modules/tree/1.0.0/packages/embroider-css-modules-temporary#api',
+  },
+);
 
-type Styles<T extends IndexSignatureParameter> = Record<
-  LocalClassName<T>,
-  string
->;
-
-type MaybeLocalClassName<T extends IndexSignatureParameter> =
-  | LocalClassName<T>[]
-  | LocalClassName<T>
-  | undefined
-  | null;
-
-interface LocalClassNewHelperSignature<T extends IndexSignatureParameter> {
-  Args: {
-    Positional: [Styles<T>, ...MaybeLocalClassName<T>[]];
-  };
-  Return: string;
-}
-
-export default class LocalClassNewHelper<
-  T extends IndexSignatureParameter,
-> extends Helper<LocalClassNewHelperSignature<T>> {
-  compute(positional: LocalClassNewHelperSignature<T>['Args']['Positional']) {
-    const [styles, ...localClassNames] = positional;
-
-    assert('The styles object is undefined.', styles);
-
-    const classNames = localClassNames.reduce<string[]>(
-      (accumulator, localClassName) => {
-        if (localClassName === undefined || localClassName === null) {
-          return accumulator;
-        }
-
-        if (Array.isArray(localClassName)) {
-          accumulator.push(...localClassName.map((element) => styles[element]));
-        } else {
-          accumulator.push(styles[localClassName]);
-        }
-
-        return accumulator;
-      },
-      [],
-    );
-
-    return classNames.filter(Boolean).join(' ');
-  }
-}
+export default LocalHelper;
