+++ packages/embroider-css-modules-temporary/src/index.ts
@@ -1,1 +1,2 @@
+export { default as local } from './helpers/local.ts';
 export { default as localClassNew } from './helpers/local-class-new.ts';
