+++ packages/embroider-css-modules-temporary/src/template-registry.ts
@@ -1,5 +1,7 @@
+import type LocalHelper from './helpers/local.ts';
 import type LocalClassNewHelper from './helpers/local-class-new.ts';
 
 export default interface EmbroiderCssModulesTemporaryRegistry {
+  local: typeof LocalHelper;
   'local-class-new': typeof LocalClassNewHelper;
 }
